feat: Enable the USB ports wakeup
O6N has a Type-C USB port connected to a PD chip using the driver
rts5453h; the port wakeup functionality is enabled using a udev rule in
`task-cd8180.typec-pd-wakeup.udev`.

The rule in `task-cd8180.usb-hid-wakeup.udev` is used to enable the
wakeup functionality of the Human Interface Devices, for example, the
mouses.

To enable the Type-A USB ports' wakeup functionality, a systemd hook is
added. The 'post' case in `cix-usb-wakeup` is necessary, because we
found that, if O6N is powered off with one of the cdns-usbssp wakeup
enabled, the O6N will boot up immediately after turning off.

Signed-off-by: Zhaoming Luo <luozhaoming@radxa.com>