fix: add a config file for masking usb wakeup
An issue was found on O6 and O6N: when the wakeup functionality of the
USB controller connected to the WiFi module was enabled before entering
S3. The machine would be woken up immediately just after it suspended.

Therefore, the config files `/usr/lib/radxa/usb-wakeup-mask/"$board"` are
added. The config files contain the ACPI IDs of the USB controllers
which should not be accessed during the execution of
`/usr/lib/systemd/system-sleep/cix-usb-wakeup`.

Signed-off-by: Zhaoming Luo <luozhaoming@radxa.com>