fix: either cix-isp-driver or cix-isp-driver-v4l2 can be used
When installing the `task-cix-isp` package, the following error will be
given:
```
Depends: cix-isp-driver but it is not installable
```

In the Cix OS, the `cix-isp-driver` package has the following content:
```
$ dpkg -L cix-isp-driver
...
/lib/modules/6.6.89-cix-build-generic/extra/armcb_isp.ko
...
```
It means it has the `armcb-isp` driver.
There's another package called `cix-isp-driver-v4l2`:
```
$ dpkg -L cix-isp-driver-v4l2
...
/lib/modules/6.6.89-cix-build-generic/extra/armcb_isp_v4l2.ko
...
```
When both `armcb-isp` and `armcb-isp-v4l2` drivers were loaded, it was
found that the V4L2 applications wouldn't work properly[0].

Therefore, we configure `cix-isp-driver` and `cix-isp-driver-v4l2` as
mutually exclusive in the `task-isp-driver` metapackage.

[0]: https://github.com/radxa-pkg/linux-sky1/commit/c1e3819a96db4f0236a36a7d83925c05b49b559c

Signed-off-by: Zhaoming Luo <luozhaoming@radxa.com>